requirements.txt
numpy>=1.24
opencv-python>=4.8
torch>=2.0
transformers>=4.30
Pillow>=9.0
